# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Endometriosis Triage & Management Application API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients/{id}/assessments:
    get:
      operationId: listPatientAssessments
      tags: [assessments]
      summary: List assessments for a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of assessments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assessment"
  /patients/{id}/investigations:
    get:
      operationId: getPatientInvestigations
      tags: [investigations]
      summary: Get investigations for a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Investigations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investigation"
    put:
      operationId: updatePatientInvestigations
      tags: [investigations]
      summary: Update or create investigations for a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestigationInput"
      responses:
        "200":
          description: Updated investigations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investigation"
  /patients/{id}/care-pathway:
    get:
      operationId: getPatientCarePathway
      tags: [care-pathway]
      summary: Get care pathway state and timeline for a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Care pathway state and events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarePathway"
    post:
      operationId: addCarePathwayEvent
      tags: [care-pathway]
      summary: Add a care pathway event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarePathwayEventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarePathwayEvent"
...
  /assessments:
    post:
      operationId: createAssessment
      tags: [assessments]
      summary: Create a new symptom assessment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentInput"
      responses:
        "201":
          description: Created assessment with triage result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assessment"
  /assessments/{id}:
    get:
      operationId: getAssessment
      tags: [assessments]
      summary: Get an assessment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assessment"
  /management-plans:
    get:
      operationId: listManagementPlans
      tags: [management-plans]
      summary: List all management plans
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: pathway
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of management plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ManagementPlan"
    post:
      operationId: createManagementPlan
      tags: [management-plans]
      summary: Create a management plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManagementPlanInput"
      responses:
        "201":
          description: Created management plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagementPlan"
  /management-plans/{id}:
    get:
      operationId: getManagementPlan
      tags: [management-plans]
      summary: Get a management plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Management plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagementPlan"
    patch:
      operationId: updateManagementPlan
      tags: [management-plans]
      summary: Update a management plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManagementPlanUpdate"
      responses:
        "200":
          description: Updated management plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagementPlan"
...
components:
  schemas:
...
    Assessment:
      type: object
      required: [id, patientId, assessmentDate, triageLevel, triageScore, createdAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        assessmentDate:
          type: string
        dysmenorrhea:
          type: integer
          description: Pain score 0-10
        chronicPelvicPain:
          type: integer
          description: Pain score 0-10
        dyspareunia:
          type: integer
          description: Pain score 0-10
        dyschezia:
          type: integer
          description: Pain score 0-10
        dysuria:
          type: integer
          description: Pain score 0-10
        infertilityHistory:
          type: boolean
        previousSurgery:
          type: boolean
        familyHistory:
          type: boolean
        symptomDurationMonths:
          type: integer
        impactOnQuality:
          type: integer
          description: 0-10 quality of life impact
        irregularBleeding:
          type: boolean
        bloating:
          type: boolean
        fatigue:
          type: boolean
        suggestedStage:
          type: ["string", "null"]
        triageLevel:
          type: string
          description: "urgent, high, moderate, routine"
        triageScore:
          type: integer
        clinicianNotes:
          type: ["string", "null"]
        createdAt:
          type: string
    AssessmentInput:
      type: object
      required: [patientId, assessmentDate, dysmenorrhea, chronicPelvicPain, dyspareunia, dyschezia, dysuria, infertilityHistory, previousSurgery, familyHistory, symptomDurationMonths, impactOnQuality, irregularBleeding, bloating, fatigue]
      properties:
        patientId:
          type: integer
        assessmentDate:
          type: string
        dysmenorrhea:
          type: integer
          minimum: 0
          maximum: 10
        chronicPelvicPain:
          type: integer
          minimum: 0
          maximum: 10
        dyspareunia:
          type: integer
          minimum: 0
          maximum: 10
        dyschezia:
          type: integer
          minimum: 0
          maximum: 10
        dysuria:
          type: integer
          minimum: 0
          maximum: 10
        infertilityHistory:
          type: boolean
        previousSurgery:
          type: boolean
        familyHistory:
          type: boolean
        symptomDurationMonths:
          type: integer
          minimum: 0
        impactOnQuality:
          type: integer
          minimum: 0
          maximum: 10
        irregularBleeding:
          type: boolean
        bloating:
          type: boolean
        fatigue:
          type: boolean
        clinicianNotes:
          type: string
    Investigation:
      type: object
      required: [id, patientId]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        tvusDate:
          type: ["string", "null"]
        tvusRequested:
          type: boolean
        tvusCompleted:
          type: boolean
        tvusFindings:
          type: ["string", "null"]
        tvusEndometrioma:
          type: boolean
        tvusEndometriomaSize:
          type: ["string", "null"]
        tvusDeepEndometriosis:
          type: boolean
        tvusDeepEndometriosisLocation:
          type: ["string", "null"]
        tvusAdenomyosis:
          type: boolean
        tvusNormal:
          type: boolean
        tvusNotes:
          type: ["string", "null"]
        mriDate:
          type: ["string", "null"]
        mriRequested:
          type: boolean
        mriCompleted:
          type: boolean
        mriFindings:
          type: ["string", "null"]
        mriDeepEndometriosis:
          type: boolean
        mriDeepEndometriosisLocation:
          type: ["string", "null"]
        mriEndometrioma:
          type: boolean
        mriEndometriomaSize:
          type: ["string", "null"]
        mriUretericInvolvement:
          type: boolean
        mriBowelInvolvement:
          type: boolean
        mriBladderInvolvement:
          type: boolean
        mriNormal:
          type: boolean
        mriNotes:
          type: ["string", "null"]
        laparoscopyDate:
          type: ["string", "null"]
        laparoscopyRequested:
          type: boolean
        laparoscopyCompleted:
          type: boolean
        laparoscopyType:
          type: ["string", "null"]
        laparoscopyFindings:
          type: ["string", "null"]
        laparoscopyRafsStage:
          type: ["string", "null"]
        laparoscopyEnzianScore:
          type: ["string", "null"]
        laparoscopyLocations:
          type: array
          items:
            type: string
        laparoscopyComplications:
          type: ["string", "null"]
        laparoscopyNotes:
          type: ["string", "null"]
        ca125Date:
          type: ["string", "null"]
        ca125Requested:
          type: boolean
        ca125Completed:
          type: boolean
        ca125Value:
          type: ["string", "null"]
        ca125Notes:
          type: ["string", "null"]
        fbcDate:
          type: ["string", "null"]
        fbcRequested:
          type: boolean
        fbcCompleted:
          type: boolean
        fbcNotes:
          type: ["string", "null"]
        requestedBy:
          type: ["string", "null"]
        reviewedBy:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    InvestigationInput:
      type: object
      properties:
        tvusDate:
          type: string
        tvusRequested:
          type: boolean
        tvusCompleted:
          type: boolean
        tvusFindings:
          type: string
        tvusEndometrioma:
          type: boolean
        tvusEndometriomaSize:
          type: string
        tvusDeepEndometriosis:
          type: boolean
        tvusDeepEndometriosisLocation:
          type: string
        tvusAdenomyosis:
          type: boolean
        tvusNormal:
          type: boolean
        tvusNotes:
          type: string
        mriDate:
          type: string
        mriRequested:
          type: boolean
        mriCompleted:
          type: boolean
        mriFindings:
          type: string
        mriDeepEndometriosis:
          type: boolean
        mriDeepEndometriosisLocation:
          type: string
        mriEndometrioma:
          type: boolean
        mriEndometriomaSize:
          type: string
        mriUretericInvolvement:
          type: boolean
        mriBowelInvolvement:
          type: boolean
        mriBladderInvolvement:
          type: boolean
        mriNormal:
          type: boolean
        mriNotes:
          type: string
        laparoscopyDate:
          type: string
        laparoscopyRequested:
          type: boolean
        laparoscopyCompleted:
          type: boolean
        laparoscopyType:
          type: string
        laparoscopyFindings:
          type: string
        laparoscopyRafsStage:
          type: string
        laparoscopyEnzianScore:
          type: string
        laparoscopyLocations:
          type: array
          items:
            type: string
        laparoscopyComplications:
          type: string
        laparoscopyNotes:
          type: string
        ca125Date:
          type: string
        ca125Requested:
          type: boolean
        ca125Completed:
          type: boolean
        ca125Value:
          type: string
        ca125Notes:
          type: string
        fbcDate:
          type: string
        fbcRequested:
          type: boolean
        fbcCompleted:
          type: boolean
        fbcNotes:
          type: string
        requestedBy:
          type: string
        reviewedBy:
          type: string
    CarePathway:
      type: object
      properties:
        state:
          $ref: "#/components/schemas/CarePathwayState"
        events:
          type: array
          items:
            $ref: "#/components/schemas/CarePathwayEvent"
...
    CarePathwayEvent:
      type: object
      required: [id, patientId, state, eventDate]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        state:
          type: string
        previousState:
          type: ["string", "null"]
        eventDate:
          type: string
        clinician:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        pathwayAssigned:
          type: ["string", "null"]
        pathwayRationale:
          type: ["string", "null"]
        surgeryId:
          type: ["integer", "null"]
        medicationName:
          type: ["string", "null"]
        medicationTier:
          type: ["string", "null"]
        waitingListType:
          type: ["string", "null"]
        targetWaitWeeks:
          type: ["integer", "null"]
        mdtSpecialties:
          type: array
          items:
            type: string
        mdtDecision:
          type: ["string", "null"]
        followUpType:
          type: ["string", "null"]
        followUpDate:
          type: ["string", "null"]
        dischargeReason:
          type: ["string", "null"]
        dischargePlan:
          type: ["string", "null"]
        createdAt:
          type: string
    CarePathwayEventInput:
      type: object
      required: [state, eventDate]
      properties:
        state:
          type: string
        eventDate:
          type: string
        clinician:
          type: string
        notes:
          type: string
        pathwayAssigned:
          type: string
        pathwayRationale:
          type: string
        surgeryId:
          type: integer
        medicationName:
          type: string
        medicationTier:
          type: string
        waitingListType:
          type: string
        targetWaitWeeks:
          type: integer
        mdtSpecialties:
          type: array
          items:
            type: string
        mdtDecision:
          type: string
        followUpType:
          type: string
        followUpDate:
          type: string
        dischargeReason:
          type: string
        dischargePlan:
          type: string
...
    ManagementPlan:
      type: object
      required: [id, patientId, status, createdAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        assessmentId:
          type: ["integer", "null"]
        status:
          type: string
          description: "active, completed, on-hold"
        approach:
          type: ["string", "null"]
          description: "medical, surgical, combined, watchful-waiting"
        pathway:
          type: ["string", "null"]
          description: "medical, surgery_general, surgery_specialist, combined, watchful_waiting"
        pathwayRationale:
          type: ["string", "null"]
        fertilityPriority:
          type: boolean
        fertilityClinicReferral:
          type: boolean
        bsgeReferral:
          type: boolean
        bsgeCentre:
          type: ["string", "null"]
        painClinicReferral:
          type: boolean
        mdtDiscussed:
          type: boolean
        mdtDate:
          type: ["string", "null"]
        mdtDecision:
          type: ["string", "null"]
        medications:
          type: array
          items:
            type: string
        surgicalOptions:
          type: array
          items:
            type: string
        lifestyleRecommendations:
          type: array
          items:
            type: string
        followUpWeeks:
          type: ["integer", "null"]
        nextReviewDate:
          type: ["string", "null"]
        goals:
          type: ["string", "null"]
        clinicianNotes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
...