# lib/api-zod/src/generated/types/carePathway.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Endometriosis Triage & Management Application API
 * OpenAPI spec version: 0.1.0
 */
import type { CarePathwayEvent } from './carePathwayEvent';
import type { CarePathwayState } from './carePathwayState';

export interface CarePathway {
  state?: CarePathwayState;
  events?: CarePathwayEvent[];
}